feat: create a resuable card component
