+++ pages/feedback.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Feedback from '../modules/assessment/feedback';
+
+const feedback = () => {
+  return (
+    <div>
+      <Feedback />
+    </div>
+  );
+};
+
+export default feedback;
